Prefer Normalized Directory Names, but Keep Existing for Compatability
PyPI proper will redirect users when they type in the name incorrectly,
such as ``/simple/DjAnGo/``, however often times bandersnatch mirrors
won't do this due to the static nature of the mirrors. Historically
this has worked because the client tools would fall back to looking
at /simple/ and figure out the link from there.

However with the changes to remove external URLs, this is having
unintended consequences where tools are no longer doing that fallback
which is "breaking" mirrors when names are not typed exactly right.
This change will normalize all names on the mirror so that clients
no longer have to guess at what the representation of a project is,
it is always the normalized form.