Avoid race condition where updates to the index page confuse clients.
The old situation is "eventually consistent" as it promises but it appears
that quite a few clients get sufficiently confused and deliver 404s instead
of falling back to a lower version.

The new approach reorders fetching the index page, the files, and providing
the updated index file to clients. Details are explained by @dstufft in #54.