Add `root_uri` configuration option.
This allows for the maintainers of the PyPI internal bandersnatch mirror which is used for failover to configure an absolute url rather than relative urls.

Specifically, warehouse will begin serving files from https://files.pythonhosted.org, so when a request fails over to the internal mirror, we want to present the same absolute URL from the bandersnatch mirror.