Remove broken delete logic
We've decided to move all deletion logic to a batch job so that we can
leave bandersnatch stateless. So this diff removes all the deletion logic
that is only working today for simple, serversig and json

This will all be moved and well unit tested in a new 'verify' option
that bandersnatch will get. This will have an optional `--delete`
that will clean up unowned files and packages with 0 remaining release files