Improve logging of the filter operations
Add an is_project_filtered() function to easily find if a project is in
a filter rule..

Added more debug logging to show when releases are getting filtered.

Fix a one character typo that was causing an exception in the release
filters because the wrong attribute name was being used.

Fixed a log message that was printing the formatting string because
the string wasn't an f-string.