Fix conf.py so Markdown docs can build properly (#557)
* Fix conf.py so Markdown docs can build properly

Since the bump to Sphinx 3.x, all of the Markdown documentation wasn't building
since the Markdown configuration was outdated. Instead of using source_parsers,
you just add the recommonmark extension. Also for good measure, Sphinx 3.0 < is
now required.

* Add protection for incompatible recommommark versions

recommonmark version 0.4.0 and below are incompatible with Sphinx 3.x and since
we moved to Sphinx 3.x, a safeguard should be added.

* Please flake8