Fix duplication of checks on internal PRs
Internal PRs (that come from branches of pypa/bandersnatch) match both
the push and pull_request events. This leads to a duplication of ci
workflows checks on such PRs. Not only is it a waste of resources, it
makes the Status Checks UI more frustrating to go through.

Patch taken from:
https://github.community/t/duplicate-checks-on-push-and-pull-request-simultaneous-event/18012