Do some good ol' CI/CD housekeeping (#933)
- Move the doc_build workflow to latest Python
- Fix some step name issues
- Merge uvloop workflow into ci workflow because their only difference
  is whether the uvloop dependency is installed. Merging them will
  avoid slow and unnecessary dependency downloads and installs and will
  get rid of one more ugly skipped job.

I also added --use-feature=in-tree-build to our CI local installs to
be good users of sister project pip. Hopefully we don't encounter
any issues, but if we do then maybe we will be able to provide some
feedback. I only added it for CI because locally because 1) we
can't always trust the local pip version supporting the flag, and 2)
avoiding too much noise when things do break.

Co-authored-by: Cooper Lees <me@cooperlees.com>