fix: draft fix for sync vulnerability
In short, the vulnerability occurs when the official mirroring and synchronization tool provided by Python Software Foundation, which is bandersnatch, failed to generate the correct Simple Index and notify the client (which is usually pip or poetry) that some packages have already been deleted by the upstream repo (PyPI.org). For example, package A was completely removed by PyPI (which means all associated releases or the project on PyPI is removed). However, third-party PyPI mirrors using bandersnatch not only still keeps the source and the wheel of A, and would still make package A downloadable and installable by client who decided to use that mirror.

Co-authored-by: BestOwl <midnightow1@outlook.com>
Co-authored-by: isnowify <isnowify@outlook.com>
Co-authored-by: TsingLoo <xiaonan.pan19@student.xjtlu.edu.cn>