Add command options for bandersnatch service in docker-compose.yml
Added multiple command options for the bandersnatch service in the docker-compose.yml file. These options include mirror, force, debug, verbose, verify, and once modes. Each mode is explained in the comments and can be uncommented to use. This provides flexibility in running the bandersnatch service based on the user's needs.