Update nginx service configuration in docker-compose.yml
This commit updates the bandersnatch_nginx service in the docker-compose.yml file. Changes include:

1. Specifying the correct build context and Dockerfile path for the Nginx image. The build context has been updated to point to the '../banderx' directory, which contains the Dockerfile and nginx.conf as per the structure. (since docker-compose doesn't handle links well in some versions)

2. Mounting the correct nginx.conf file.
Previously, the nginx.conf file was assumed to be in the current directory. This update mounts the nginx.conf from '../banderx', following the instructions in '../banderx/README.md', ensuring that the Nginx server uses the appropriate configuration file for serving the mirrored data.

3. Updated comment has been updated to more accurately describe its purpose - serving the downloaded repository.