Refactoring of Storage and Bander...Mirror inits
Change the Storage base class and the storage plugin loading functions
to require a ConfigParser instance so the BandersnatchConfig singleton
is not accessed.

Change the BandersnatchMirror class to require a Storage instance
instead of a storage backend name since the mirror subcommand already
loads a storage plugin instance.

The diff in the application code is relatively simple, but both changes
caused cascading/rippling updates to the unit tests that will require a
regrettable amount of tedious review. It would be nice to find a way to
split up this commit.