Refactoring of Filter & Mirror initialization
Change the Filter base class and filter loader to require an external
ConfigParser instance be passed in.

Change the Mirror/BandersnatchMirror classes to require an external
instance of LoadedFilters.

op/ed: Remember all that nonsense you told yourself about removing a
singleton making the tests simpler? Well so far that is not the case.