Change delete_path to use asyncio's current event loop
delete.py::delete_path was using the event loop saved on the current storage plugin,
i.e. storage_backend.loop, for running storage plugin functions asynchronously.

With pytest >8.x, the corresponding test test_delete.py::test_delete_path began failing
because the loop attached to the storage plugin was closed when the function tried to
use it. This is presumably due to changes to pytest and pytest-asyncio. It behaves like
the event loop is set when the storage plugin is initialized, then between tests that
loop is closed and a new one is created, leaving the plugin holding the closed instance.
If a test then accessed the cached instance of the storage plugin, it would in turn get
the now-invalid event loop instance.

Other uses of the run_in_executor + storage_backend pattern seem to have avoided this
problem  because they either call 'storage_backend_plugins' with it's clear_cache
argument, or were already calling asyncio.get_event_loop or equivalent.

When Bandersnatch is actually running, there will only be one event loop for the
lifetime of the program so the value of storage_backend.loop and asyncio.get_event_loop
will be the same instance, and this change will make no difference.