Reset config singleton for all tests in test_main.py
Fixes order-dependent test failures in test_main.py caused by tests that
tried to initialize the singleton BandersnatchConfig instance when it
had already been created. This left the config out of sync with what the
tests expected.

Some tests in this module were already doing this with a call to a
'setup' function that erased the Singleton metaclass's stored instances.
This moves that into a pytest fixture and applies it to the whole module.