Set object store URL in tests using an env var
The S3 endpoint URL used in the unit tests was hard-coded to localhost.
When executing the tests in a container, a minio sibling container can't
be accessed via localhost. This allows setting the endpoint URL via an
environment variable, and adds 'BANDERSNATCH_*' to the environment
variable passthrough list for Tox.

...noticing that Tox was the reason the environment variable was being
ignored took a very long time.