Fix line endings for storage plugin sample file test
Previously the test_get_hash test used shutil.copy to duplicate a sample file
from the project directory into a temp directory for the test.

If the project folder is cloned on Windows then bind-mounted into
a Linux container, with the test run in the container, the file
would have Windows line endings while the test would expect unix
line endings based on sys.platform within the container.

Instead of copying the file, this change writes the file from a string
value using Python's universal newlines support, so the written file
will have the same line ending as the platform executing the test.

This change trips mypy errors in the pre-commit checks. The errors show
when running mypy in the console or via pre-commit, but the MyPy plugin
for VS Code *doesn't* and will make the *opposite complaint* if you try
to change the file to fix the errors shown in pre-commit. I don't know
what to do about that.