Set some pytest file discovery settings explicitly in pytest.ini
Pytest would fail to discover any test files when run in a container
with the project folder mounted. Adding these helped, and doesn't
change anything when tests are run 'normally' as far as I can tell.

I know what each does, but I don't know why the problem was happening
in the container specifically, or why these fix it. This is a "web search
the error message" -> "find reasonable option on StackOverflow" fix.