Upgrade test dependencies pytest and pytest-asyncio  (#1829)
* Manually update pytest and pytest-asyncio in requirements_test.txt

This upgrades the two packages together:
- pytest => 8.x
- pytest-asyncio => 0.24.x

Commands used:

    pip install -U pytest pytest-asyncio
    pip freeze -r requirements_test.txt

* Add recommended 'asyncio_default_fixture_loop_scope' to pytest.ini

0.24.x of pytest-asyncio adds a deprecation warning for the default
event loop fixture scope changing from fixture-cache to per-function.

This pre-emptively sets this option to align with the upcoming default.

* Reset config singleton for all tests in test_main.py

Fixes order-dependent test failures in test_main.py caused by tests that
tried to initialize the singleton BandersnatchConfig instance when it
had already been created. This left the config out of sync with what the
tests expected.

Some tests in this module were already doing this with a call to a
'setup' function that erased the Singleton metaclass's stored instances.
This moves that into a pytest fixture and applies it to the whole module.

* Add pytest-asyncio marks to tests that implicitly require an event loop

Tests that are not coroutines themselves may require an event loop to
exist for code that uses get_event_loop/get_running_loop.

* Clear cached storage backends between test modules

Adds an autouse fixture at the module scope that removes all loaded
storage plugins. This prevents plugin instances with closed event loops
from being re-used across tests.

Also adds a non-autouse fixture for resetting the storage plugins at
more narrow scopes when required.

* Remove a left behind event loop reassignment from a test